Save state

+++ db/migrate/20251115235847_remove_area_from_disciplina.rb
@@ -1,5 +1,5 @@
 class RemoveAreaFromDisciplina < ActiveRecord::Migration[7.1]
   def change
-    remove_column :disciplinas, :area, :string
+    remove_column :disciplinas, :area, :string if column_exists?(:disciplinas, :area)
   end
 end
